+++ e310x/src/qspi0/delay1.rs
@@ -25,12 +25,12 @@ impl R {
 impl W {
     #[doc = "Bits 0:7 - Minimum CS inactive time"]
     #[inline(always)]
-    pub fn intercs(&mut self) -> IntercsW<Delay1Spec> {
+    pub fn intercs(&mut self) -> IntercsW<'_, Delay1Spec> {
         IntercsW::new(self, 0)
     }
     #[doc = "Bits 16:23 - Maximum interframe delay"]
     #[inline(always)]
-    pub fn interxfr(&mut self) -> InterxfrW<Delay1Spec> {
+    pub fn interxfr(&mut self) -> InterxfrW<'_, Delay1Spec> {
         InterxfrW::new(self, 16)
     }
 }