+++ e310x/update.sh
@@ -2,11 +2,10 @@
 set -x
 set -e
 
-# used svd2rust 0.34.0
+# used svd2rust 0.36.1
 rm -rf src
 mkdir src
-../../svd2rust/target/release/svd2rust --target riscv --settings settings.yaml -g -i e310x.svd
-# svd2rust --target riscv --settings settings.yaml -g -i e310x.svd
+svd2rust --target riscv --settings settings.yaml --feature-group -g -i e310x.svd
 rm -f features.toml
 mv generic.rs src/
 form -i lib.rs -o src/